+++ Library/Allocators.c
@@ -21,7 +21,7 @@ void k_Containers_Teardown()
 	allocators = NULL;
 }
 
-k_Allocator *k_GetAllocator(int elementSize)
+k_Allocator *k_GetAllocator(size_t elementSize)
 {
 	for (int n = 0; n < allocators->size; ++n)
 	{
